backend/pyproject.toml
[project]
name = "agent-studio-backend"
version = "0.1.0"
description = "Local-first Agent Runner + tracing store for Agent Studio (FastAPI + SQLite)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "pydantic-settings>=2.2.1",
  "orjson>=3.10.0",
]

[project.optional-dependencies]
agents = [
  "openai-agents>=0.2.0",
]
sandbox = [
  "playwright>=1.40.0",
  "pyseccomp>=0.1.2;platform_system=='Linux'",
]

[project.scripts]
agent-studio-backend = "agent_studio_backend.main:run"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_studio_backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

